sync

+++ qsandbox_default/code/qcommon/q_shared.h
@@ -1198,7 +1198,7 @@ typedef struct usercmd_s {
 	int				serverTime;
 	int				angles[3];
 	int 			buttons;
-	byte			weapon;           // weapon
+	int				weapon;           // weapon
 	signed char	forwardmove, rightmove, upmove;
 } usercmd_t;
